Parametrize script for generating secrets
With this change, generate-secrets.sh can be passed the "-f" or
"--force" parameter, which will not prompt the user when generating
secrets.

Additionally, secret generation can be prevented for individual secrets
by setting an appropriate environment variable (JORE4HASURA_XXX).

These changes serve the use of the script in integration tests. This is
why the script was also renamed.