Set CASCADE on infra link and stop point vehicle_(sub)mode refs
The information contained by the vehicle_submode_on_infrastructure_link
and vehicle_mode_on_scheduled_stop_point tables should be interpreted to
hold information on the infrastructure_link and scheduled_stop_point
entities, respectively.

For this reason, it is safe to cascade those tables' affected rows'
information when the entitity referencing those rows is changed such
that the reference is broken.