Scheduled stop point infra link constraints
Constraints to enforce that the stop points of a journey pattern can
only reside on infra links included in the route of that journey
pattern. Additionally, their order must be the same as the order of the
infra links they reside on and they must be stopped at in a valid
direction.