Prevent multiple journey patterns under the same route
At the moment, the UI supports only a single journey pattern under the
same route. Yet, we want to maintain the possibility to add multiple
journey patterns under the same route later.

Therefore, we leave the data model structure as is and restrict the
number of journey patterns to 1 by creating a unique index on the
journey_pattern table.

This change can later be undone when the UI supports this. Simply create
a new commit, adding a new migration containing this commit's up.sql and
down.sql with swapped content (rename up.sql to down.sql and vice
versa).