Optimize route<->journey pattern verification constraints
Previously, the route<->journey pattern consistency check was run
separately for every significant changed row in any of the related
tables:
  - journey_pattern.scheduled_stop_point_in_journey_pattern
  - route.infrastructure_link_along_route
  - internal_service_pattern.scheduled_stop_point
  - journey_pattern.journey_pattern

This was realized via deferred row-level constraint triggers only and
thus caused a lot of overhead.

With this change, we create statement-level triggers for the significant
operations on above mentioned tables to enqueue verification of the
changed journey patterns and routes using a temp table. Then we use
deferred constraint triggers at the end of transactions to check the
consistency of each queued journey pattern / route only once.