Fix route verification check
...with regards to multiple stops on the same infra link.

Before this fix, multiple stop instances with the same label could in
some situations be placed in different stop order positions if the
conflicting stop with another label was on the same link as one of the
stop instances in question.

For example this illegal situation did not get caught by the check:

  infra links:  -----+-------+-------+------
  stops:          A      B      C  B     D

This was due to a missing "order by" column spec, which would ensure
that the second instance of stop B would be in the correct order with
stop C.

This "order by" column spec was added in this commit.