Priority based route / journey pattern verification
Previously, all instances of a route, which were valid at the same time,
had to have journey patterns, which were compatible between all of the
route instances. This was due to the fact that the route / journey
pattern validity check was performed at once for all routes / journey
patterns valid at a given point in time.

With this change, only those route and scheduled stop point instances
are taken into account, which have highest priority at a given point in
time. This allows for:

1. Temporary routes' geometries to deviate from the basic route
instance's geometry, while including temporary stop instances in the
journey pattern, which are located on links only present in the
temporary route instance (but not in the basic route instance).

2. The stop order of a temporary route to be different compared to that
of it's basic route. I.e. a temporary stop instance can be placed at a
position in the temporary route's journey pattern, which would
(according to the order of the infra links) occur only after the next
stop in the basic route's journey pattern.