Perform route verification of multiple routes in one go
The previous version of the route verification triggered individual runs
of the route verification function for every route to be checked, which
was quite slow.

This version checks all routes in one go and additionally also checks
that there are no so-called "ghost stop points", whose validity span
does not overlap with the route's validity span at all.

As a side effect, this commit also fixes a bug, which prevented complete
routes (with links and journey pattern) from being removed at once.