Integration test to verify a route with "ghost stops"
A "ghost stop" is an ssp instance on a journey pattern, whose validity
span does not overlap with the route's validity span at all. This should
be prevented, because the stop will never actually appear on the journey
pattern. (Unless its validity period is changed and it might appear
without the user being aware of it.)