Tests: Functionality for disabling triggers when setting up db
For unit tests, we do not require that the database is set up in a way
in a way satisfying all constraints. Usually it is enough to provide
the mandatory input data to the unit tested.

Therefore, this commit provides functionality for disabling triggers
when setting up the database for testing.