Generate database dumps for comparability
The /scripts/generate-dumps.sh script takes a database dump from both the routes and timetables databases. This is then copied to the migrations folder.
These dumps are to help with code reviews: from the dump it's easier to see the actual diff to the database schema compared to the verbose migration scripts