Fix insert_scheduled_stop_point_with_vehicle_mode function
The function had to be redefined in the validity_time_as_date migration.
The new definition inserted the scheduled stop point into the
internal_service_pattern.scheduled_stop_point-table, skipping the
service_pattern.scheduled_stop_point-view's trigger, which takes care of
inserting a new scheduled_stop_point_invariant when needed. This commit
re-defines the function the correct way.

The function was not previously tested, since it is not exposed in the
gql-API. This commit also adds rudimentary unit-tests for the function.