Start using the hsl datamodel for integration tests
The routes module had not have anything to be tested outside of the generic model, thus the integration tests scoped the generic model as well instead of the HSL one. The timetables module has much more non-Transmodel, HSL-exclusive model which should also be tested with integration tests. Thus, this commit aims to make the integration tests compatible with the HSL datamodel.