Fix test data inserting so tests can be ran again
This broke after previous commit: didn't work with computed columns.

Earlier approach was basically "casting" the JSON dataset to DB format
with jsonb_populate_recordset, and SELECTing that for insert.
Problem was, the SELECT would return the computed column as well,
which cannot be included in insert.
Tried eg. "select everything but the computed column" for insert
(not possible), but in the end had to resort to this hacky approach.
Basically, use INSERT VALUES syntax instead, and build the values list manually
based on the JSON test data objects.

A better solution might be to use knex instead, might try that later.