Refactor drop trigger/constraint/function code blocks to functions
So we can invoke them from hsl side as well,
without polluting generic migrations with hsl tables.

Also tried implementing this with functions that would be called
just for a single schema at a time,
but that would not work because of dependencies between the dropped entities.
Cascade deletes would have worked, but this feels safer (and closer to previous approach).