Make legacyHslMunicipalityCode nullable
Because of course if it is NOT NULL we couldn't create new routes/lines
from Jore 4 UI until it can set the field. Which it can't currently,
and not sure yet if it even should.
So, make this nullable until we specify how this should be handled
for new routes/lines.