Prevent multiple passing times in vehicle journey from referencing same instance of scheduled stop point
Mostly for cases when these are one after another in the sequence.
If they are spread out, then other conditions would already fail.

Could have implemented this in the validation function as well,
but felt that index is more suitable for enforcing this.
Of course, if it looks like we have problems with index performance / size
then this can be revisited.