Add timetable version return value table
This introduces the schema return_value which is used to have all the
tables that will not contain any data, but are used for function return
values as they are required to have. There are already this type of
return value tables in the database that are not located in this scheme,
but they need to be moved under this schema in the future. Although
this commit will add timetable_version as the first return_value.