Eliminate duplicate values from recursive query parameters
Seemed to radically worsen the resucrsive query performance
if this happened.
Eg. in one insert the sspijp_ilar_combos contained 45 rows,
which turned out to be 15 actual ones + 2x duplicates for each.
This resulted in the resursive query (traversal)
returning around 21.5 million rows and taking an hour.
After eliminating the duplicates it returns just 15 rows
and takes 20ms.

Another way would be to try to fix this in the recursive
query itself, either by preventing loops
or preventing joins to duplicate rows,
but didn't feel straightforward (eg. I don't think
SSP id is safe because this is used from UPDATE triggers).