Fix internal utils migrations
There were an accidental change made to an old migration which was not an repeatable migration
this caused the dev and test environments to crash, because the changes did not obviously come
in to action. So in this commit we change the old migration like it was and create a new migration
for these changes. Immutable functions can't be in repeatable migrations.