Add new versions of vehicle journey end_time and start_time functions
We have these functions that read intervals as text and returns them as text, because
of some reason which blocked hasura from using the interval-versions from being used
as calculated fields in a table. There is a comment about this on those functions in this same file.
But if we want to use these internally in db's other functions, we can use the interval versions.
These will be used in upcoming get_vehicle_schedules_on_date function.