Add helpers for accessing built dataset
With deeply nested objects accessing them is quite verbose, eg:
  builtDataset
    ._vehicle_schedule_frames.winter2022
    ._vehicle_services.sun
    ._blocks.block
    ._vehicle_journeys.route123Inbound
    .vehicle_journey_id

It is quite common that we have unique label for each row.
Thus, with these helpers, one can access eg. vehicle journeys like this:
  const datasetHelper = createTimetablesDatasetHelper(builtDataset);
  datasetHelper.getVehicleJourney('route123Inbound');

Of course, this only works reliably if the labels are actually unique:
it could be that there are 'route123Inbound' used in multiple blocks.
In that case an error is thrown, forcing the developer to use a more robust
way and not create flaky / incorrect assertions.