Add substitute period validity range to schedule calculations
This was two birds with one stone, first of all the initial vsf.validity_start and vsf.validity_end were
wrong, and caused the substitute operating day to show the validity of the target day instead
of the superseded_date. But also now we have the substitute operating period  implemented
and those can last multiple days and that is the real validity range.
 Also added the used internal_utils functions to calculate the begin and end date here.

Resolves HSLdevcom/jore4#1390