Fix Luxon type inferring when using the built version of data inserter
Ran into some problems when trying to include this in a rollup build.
These instanceof calls were not working, probably because
the Luxon constructor was not the same instance on TDI and caller side.
Using flags present in the Luxon objects instead.
Slight problem with these is that they are marked as @private
and not present in the TS definitions, thus we need some type casting.