Add hardcoded ids to vehicle types populated in migrations
These missing meant that vehicle types got a random UUID each time.
This makes creating test data a bit more difficult sometimes,
because for blocks the vehicle_type_id is a required value.
One example of this is the test-db-manager seed script, which has to
fetch vehicle types from a running DB before it can create the seed data.
If we hard code the ids, we can just use known values in such test code.

We already do this kind of hard coded UUIDs for values populated in migrations
at least for service_calendar.day_type and service_calendar.day_type_active_on_day_of_week tables.

This only affects new databases and not eg. our test (demo) environment,
but that is fine, since this fix is meant to address issues in test data
and the id values are not relied upon in any app code.

There didn't seem to be any other tables that had a similar issue.