Add stop_place_newest_version metadata
Add the necessary metadata to track stop_place_newest_version which
is a view that only has the newest versions of a stop (distincted by netex_id).
Also add the necessary relationships, a remote relationship to tiamat API
stop place and remote relationship to scheduled_stop_point. We might have
use for both, or only one of these remote relationships, but should be
no harm in adding both.

Resolves https://dev.azure.com/hslfi/JORE%204.0/_workitems/edit/35459