Add join to newest StopPlace from newest Quay
Provides same data as the more direct join to stop_place table, but as
the stop_place table and stop_place_newest_version view are not the
same DB table/view they are treated as 2 completely different types in
the Graph. Thus, reusing code (fragments) between the 2 is not
possible, and by default the UI code uses the stop_place_newest_version
view for queries -> add a join to that.